Deleting Files

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require("express-validator/check");
+const fileHelper = require('../util/file')
 
 const Product = require("../models/product");
 
@@ -135,6 +136,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl)
         product.imageUrl = image.path;
       }
       return product.save().then((result) => {
@@ -221,14 +223,22 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log("DESTROYED PRODUCT");
-      res.redirect("/admin/products");
-    })
-    .catch((err) => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    }); 
+  Product.findById(prodId)
+  .then(product => {
+    if (!product) {
+      return next(new Error('Product not found.'))
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ _id: prodId, userId: req.user._id });
+  })
+  .then(() => {
+    console.log("DESTROYED PRODUCT");
+    res.redirect("/admin/products");
+  })
+  .catch((err) => {
+    console.log(err)
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
+  }); 
 };
